requirements.txt
google-genai
pdf2image
moviepy
Pillow
python-dotenv
